+++ packages/Uploader/src/components/File/File.js
@@ -2,7 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import CheckIcon from "@paprika/icon/lib/Check";
 import TimesIcon from "@paprika/icon/lib/Times";
-import RetryIcon from "@paprika/icon/lib/Refresh";
+import InfoCircle from "@paprika/icon/lib/InfoCircle";
+
 import Button from "@paprika/button";
 import useI18n from "@paprika/l10n/lib/useI18n";
 import Popover from "@paprika/popover";
@@ -23,6 +24,7 @@ const propTypes = {
   error: PropTypes.string,
   fileKey: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
+  onError: PropTypes.func,
   progress: PropTypes.number,
   size: PropTypes.number.isRequired,
   status: PropTypes.oneOf(Object.keys(File.types.status).map(key => File.types.status[key])).isRequired, // eslint-disable-line no-use-before-define
@@ -30,11 +32,14 @@ const propTypes = {
 
 const defaultProps = {
   error: "",
+  onError: null,
   progress: 0,
 };
 
-function File({ error, fileKey, name, progress, size, status }) {
-  const { cancelFile, restartFileUpload } = React.useContext(UploaderContext);
+function File(props) {
+  const { error, fileKey, name, progress, size, status, onError } = props;
+
+  const { cancelFile } = React.useContext(UploaderContext);
   const I18n = useI18n();
   const sizeWithUnits = getNumberWithUnits(I18n, size);
   const progressWithUnits = getNumberWithUnits(I18n, (size * progress) / 100);
@@ -47,14 +52,8 @@ function File({ error, fileKey, name, progress, size, status }) {
           <Popover isDark isEager>
             <Popover.Tip />
             <Popover.Trigger>
-              <Button.Icon
-                kind={Button.Icon.types.kind.MINOR}
-                onClick={() => {
-                  restartFileUpload(fileKey);
-                }}
-                size={Button.Icon.types.size.SMALL}
-              >
-                <RetryIcon />
+              <Button.Icon kind={Button.Icon.types.kind.MINOR} onClick={() => {}}>
+                <InfoCircle />
               </Button.Icon>
             </Popover.Trigger>
             <Popover.Content>
@@ -90,7 +89,7 @@ function File({ error, fileKey, name, progress, size, status }) {
   function getProgressText() {
     switch (status) {
       case types.status.ERROR:
-        return error;
+        return typeof onError === "function" ? onError(error) : error;
       case types.status.SUCCESS:
         return I18n.t("uploader.progress.complete");
       case types.status.CANCEL: