+++ packages/Uploader/src/useProcessFiles.js
@@ -1,5 +1,4 @@
 import React from "react";
-import uuidv4 from "uuid/v4";
 import * as types from "./types";
 import { uploadToServer } from "./helpers";
 
@@ -93,27 +92,6 @@ export default function useProcessFiles({
     }
   }
 
-  function restartFileUpload(key) {
-    const index = getFileByIndex(key, files);
-
-    if (index !== null) {
-      const file = files[index];
-
-      if (file.status === types.status.ERROR || file.status === types.status.CANCEL) {
-        setFiles(
-          setFile(file, fileItem => {
-            const file = fileItem;
-            file.key = uuidv4(); // change its key so it will be restartable
-            file.progress = 0;
-            file.request._data = undefined; // so superagent allows the upload to restart
-            file.request._aborted = false; // so superagent allows the upload to restart
-            return fileItem;
-          })
-        );
-      }
-    }
-  }
-
   const upload = React.useCallback(() => {
     const isSameList = (files, uploadingFileList) => {
       const sameList =
@@ -201,5 +179,5 @@ export default function useProcessFiles({
     }
   }, [files, hasAutoUpload, upload]);
 
-  return { files, setFiles, isDisabled, isCompleted, upload, removeFile, cancelFile, restartFileUpload };
+  return { files, setFiles, isDisabled, isCompleted, upload, removeFile, cancelFile };
 }